Remove useless abstraction
