+++ Vizir/src/Vizir/Core/MouseButtonCode.h
@@ -1,15 +1,15 @@
 #pragma once
 
 // From glfw3.h
-#define HZ_MOUSE_BUTTON_1         0
-#define HZ_MOUSE_BUTTON_2         1
-#define HZ_MOUSE_BUTTON_3         2
-#define HZ_MOUSE_BUTTON_4         3
-#define HZ_MOUSE_BUTTON_5         4
-#define HZ_MOUSE_BUTTON_6         5
-#define HZ_MOUSE_BUTTON_7         6
-#define HZ_MOUSE_BUTTON_8         7
-#define HZ_MOUSE_BUTTON_LAST      HZ_MOUSE_BUTTON_8
-#define HZ_MOUSE_BUTTON_LEFT      HZ_MOUSE_BUTTON_1
-#define HZ_MOUSE_BUTTON_RIGHT     HZ_MOUSE_BUTTON_2
-#define HZ_MOUSE_BUTTON_MIDDLE    HZ_MOUSE_BUTTON_3
+#define VZ_MOUSE_BUTTON_1         0
+#define VZ_MOUSE_BUTTON_2         1
+#define VZ_MOUSE_BUTTON_3         2
+#define VZ_MOUSE_BUTTON_4         3
+#define VZ_MOUSE_BUTTON_5         4
+#define VZ_MOUSE_BUTTON_6         5
+#define VZ_MOUSE_BUTTON_7         6
+#define VZ_MOUSE_BUTTON_8         7
+#define VZ_MOUSE_BUTTON_LAST      VZ_MOUSE_BUTTON_8
+#define VZ_MOUSE_BUTTON_LEFT      VZ_MOUSE_BUTTON_1
+#define VZ_MOUSE_BUTTON_RIGHT     VZ_MOUSE_BUTTON_2
+#define VZ_MOUSE_BUTTON_MIDDLE    VZ_MOUSE_BUTTON_3
