--- Vizir/src/Vizir/Platform/Windows/WindowsInput.h
@@ -1,17 +1,0 @@
-#pragma once
-
-#include "Core/Input.h"
-
-namespace Vizir
-{
-	class VZ_API WindowsInput : public Input
-	{
-	protected:
-		virtual bool IsKeyPressedImpl(int keycode) override;
-
-		virtual bool IsMouseButtonPressedImpl(int button) override;
-		virtual std::pair<float, float> GetMousePosImpl() override;
-		virtual float GetMouseXImpl() override;
-		virtual float GetMouseYImpl() override;
-	};
-}
