--- Vizir/src/Vizir/Platform/Windows/WindowsWindow.h
@@ -1,46 +1,0 @@
-#pragma once
-
-#include "Core/Window.h"
-#include "Renderer/GraphicsContext.h"
-
-#include <GLFW/glfw3.h>
-
-namespace Vizir
-{
-	class VZ_API WindowsWindow : public Window
-	{
-	public:
-		WindowsWindow(const WindowProps& props);
-		virtual ~WindowsWindow();
-
-		void OnUpdate() override;
-
-		inline unsigned int GetWidth() const override { return m_Data.Width; }
-		inline unsigned int GetHeight() const override { return m_Data.Height; }
-
-		inline virtual void* GetNativeWindow() const override { return m_Window; }
-
-		inline void SetEventCallback(const EventCallbackFn& callback) override { m_Data.EventCallback = callback; }
-		void SetVSync(bool enabled) override;
-		bool IsVSync() const override;
-
-	private:
-		virtual void Init(const WindowProps& props);
-		virtual void Shutdown();
-	private:
-		GLFWwindow* m_Window;
-		Scope<GraphicsContext> m_Context;
-
-		struct WindowData
-		{
-			std::string Title;
-			unsigned int Width, Height;
-			bool VSync;
-
-			EventCallbackFn EventCallback;
-		};
-
-		WindowData m_Data;
-	};
-}
-
